chore(*): Update dependencies, add rollup config
